+++ NEWS.md
@@ -21,6 +21,7 @@ Released to CRAN on 2022-10-17
 * Fix bug that prevented publishing or writing manifests for non-Quarto content
   when a Quarto path was provided to the `quarto` argument of `writeManifest()`,
   `deployApp()`, and related functions.
+* Update company and product name for rebranding to Posit
 
 * Escape account names when performing a directory search to determine an
   appropriate server. (#620)