+++ NEWS.md
@@ -4,8 +4,8 @@
 
 * Introduced support for publishing to Posit Cloud. This feature is currently
   in closed beta and requires access to an enabled account on Posit Cloud.
-  See [Posit Cloud's Announcement](https://posit.cloud/learn/whats-new) for
-  more information and to request access.
+  See [Posit Cloud's Announcement](https://posit.cloud/learn/whats-new#publishing)
+  for more information and to request access.
 
 * Update company and product name for rebranding to Posit.