+++ NEWS.md
@@ -1,5 +1,10 @@
 # rsconnect 0.8.29 (development version)
 
+* Introduced support for publishing to Posit Cloud. This feature is currently
+  in closed beta and requires access to an enabled account on Posit Cloud.
+  See [Posit Cloud's Announcement](https://posit.cloud/learn/whats-new#publishing)
+  for more information and to request access.
+
 * Update company and product names for rebranding to Posit.
 
 # rsconnect 0.8.28